+++ src/pages/index.js
@@ -1,0 +1,5 @@
+export { ProductPage } from "./ProductPage"
+export { Cart } from "./private/Cart";
+export { WishList } from "./private/Wishlist";
+export { Home } from "./Home";
+export { Products } from "../pages/AllProductsPage";
